backend/requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
pycparser==2.21
pymongo==4.0.2
pyparsing==3.0.7
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.3
